pyproject.toml
[project]
name = "laundry-monitor"
version = "0.1.0"
description = "A simple web application for managing the washing machines in the univerysity laundry room."
authors = [
    {name = "Gleb Popov",email = "g.popov@innopolis.university"},
    {name = "Aziz Vundirov",email = "a.vundirov@innopolis.university"},
    {name = "Elizaveta Zagurskih",email = "e.zagurskih@innopolis.university"},
    {name = "Aleksey Antipov",email = "a.antipov@innopolis.university"},
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "mutmut (>=3.5.0,<4.0.0)",
    "hypothesis (>=6.151.10,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)",
    "radon (>=6.0.1,<7.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "bandit (>=1.9.4,<2.0.0)",
    "locust (>=2.43.3,<3.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "xenon (>=0.9.3,<0.10.0)",
]

[tool.mutmut]
paths_to_mutate = ["src/service/machine.py"]
tests_dir = ["tests"] 
also_copy = ["src"]
max_stack_depth = 20
